Add files via upload
Hey everyone, figured I'd update the UML and upload it here; we've added several new classes and whatnot. 

The flowchart still seems mostly relevant, I think that can wait to update until we have a better idea of what we intend to do gameplay-wise. Uploading it here as well so we have everything in one place.