Upload PlayerStats.java
Created a class to hold health total. Not sure if we need it or if we'll have some other way of tracking defeat or combat results, but for now it seemed needed. Named it "Stats" in case we wanted to add something else to it, otherwise we can change it to PlayerHealth to be more accurate later.