Update Deck.java
Made the toString prettier and gave it a card number; will need that if we have the game let players choose cards.

Altered the constructors to create a default deck of size 10, and then to let someone specify the size. Now a deck will be generated right away. Card.java has been modified to support this, update going up shortly.