Update Card.java
Added attribute and name generators. Cleaned up the toString so all the prints are in a vertical line (everything isn't staggered). 

Attribute and Name generators are kind of placeholders; we have an Enum for the former and can make one for the latter, but I'm not yet sure how to randomize from an Enum, so we can fix that later.